+++ src/Button/index.js
@@ -1,208 +1,19 @@
 // @flow
 import * as React from "react";
-import styled, { css } from "styled-components";
-import { warning } from "@adeira/js";
 
-import defaultTheme from "../defaultTheme";
 import { ICON_SIZES } from "../Icon/consts";
-import { TYPE_OPTIONS, SIZE_OPTIONS, TOKENS, BUTTON_STATES } from "./consts";
-import Loading, { StyledSpinner } from "../Loading";
-import { getSize } from "../Icon";
-import getSpacingToken from "../common/getSpacingToken";
-import getSizeToken from "./helpers/getSizeToken";
-import getTypeToken from "./helpers/getTypeToken";
-import getButtonSpacing from "./helpers/getButtonSpacing";
-import getIconSpacing from "./helpers/getIconSpacing";
-import getButtonBoxShadow from "./helpers/getButtonBoxShadow";
-import getFocus from "./helpers/getFocus";
-import onlyIE from "../utils/onlyIE";
+import { TYPE_OPTIONS, SIZE_OPTIONS } from "../primitives/ButtonPrimitive/consts";
+import Loading from "../Loading";
+import ButtonPrimitive from "../primitives/ButtonPrimitive";
+import IconContainer from "../primitives/ButtonPrimitive/components/IconContainer";
+import StyledButtonContent from "../primitives/ButtonPrimitive/components/ButtonContent";
+import StyledButtonContentChildren from "../primitives/ButtonPrimitive/components/ButtonContentChildren";
 
 import type { Props } from "./index";
 
-const IconContainer = styled(({ className, children }) => (
-  <span className={className}>{children}</span>
-))`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  margin: ${getIconSpacing()};
-  color: ${({ bordered }) =>
-    bordered ? getTypeToken(TOKENS.colorTextButtonBordered) : getTypeToken(TOKENS.colorTextButton)};
-  transition: background ${({ theme }) => theme.orbit.durationFast} ease-in-out,
-    box-shadow ${({ theme }) => theme.orbit.durationFast} ease-in-out;
-
-  > svg {
-    width: ${({ sizeIcon }) => getSize(sizeIcon)};
-    height: ${({ sizeIcon }) => getSize(sizeIcon)};
-  }
-`;
-
-IconContainer.defaultProps = {
-  theme: defaultTheme,
-};
-
-export const StyledButton = styled(
-  ({
-    theme,
-    asComponent,
-    circled,
-    external,
-    type,
-    icon,
-    iconLeft,
-    iconRight,
-    sizeIcon,
-    width,
-    bordered,
-    loading,
-    onlyIcon,
-    fullWidth,
-    style,
-    dataTest,
-    submit,
-    buttonRef,
-    ariaControls,
-    ariaExpanded,
-    spaceAfter,
-    title,
-    ...props
-  }) => {
-    const isButtonWithHref = asComponent === "button" && props.href;
-    const Component = isButtonWithHref ? "a" : asComponent;
-    const buttonType = submit ? "submit" : "button";
-    return (
-      <Component
-        data-test={dataTest}
-        aria-controls={ariaControls}
-        aria-expanded={ariaExpanded}
-        aria-label={title}
-        type={!isButtonWithHref ? buttonType : undefined}
-        {...props}
-        ref={buttonRef}
-      >
-        {props.children}
-      </Component>
-    );
-  },
-)`
-  position: relative;
-  display: ${({ href, asComponent }) => (href || asComponent === "a" ? "inline-flex" : "flex")};
-  justify-content: center;
-  align-items: center;
-  box-sizing: border-box;
-  appearance: none;
-  text-decoration: none;
-  width: ${({ fullWidth, width, onlyIcon }) =>
-    fullWidth
-      ? "100%"
-      : (width && `${width}px`) || (onlyIcon && getSizeToken(TOKENS.heightButton)) || "auto"};
-  flex: ${({ fullWidth }) => (fullWidth ? "1 1 auto" : "0 0 auto")};
-  max-width: 100%; // to ensure that Buttons content wraps in IE
-  height: ${getSizeToken(TOKENS.heightButton)};
-  background: ${({ bordered }) =>
-    bordered
-      ? getTypeToken(TOKENS.backgroundButtonBordered)
-      : getTypeToken(TOKENS.backgroundButton)};
-  color: ${({ bordered }) =>
-    bordered
-      ? getTypeToken(TOKENS.colorTextButtonBordered)
-      : getTypeToken(TOKENS.colorTextButton)} !important;
-  border: 0;
-  border-radius: ${({ theme, circled }) =>
-    circled ? getSizeToken(TOKENS.heightButton) : theme.orbit.borderRadiusNormal};
-  min-width: ${({ circled }) => circled && getSizeToken(TOKENS.heightButton)}};
-  padding: ${getButtonSpacing()};
-  font-family: ${({ theme }) => theme.orbit.fontFamily};
-  font-weight: ${({ theme }) => theme.orbit.fontWeightBold}!important;
-  font-size: ${getSizeToken(TOKENS.fontSizeButton)};
-  line-height: 1.4; // preventing inheriting with safe value
-  cursor: ${({ disabled }) => (disabled ? "not-allowed" : "pointer")};
-  transition: all 0.15s ease-in-out !important;
-  outline: 0;
-  opacity: ${({ disabled, theme }) => disabled && theme.orbit.opacityButtonDisabled};
-  margin-bottom: ${getSpacingToken};
-  ${getButtonBoxShadow(BUTTON_STATES.DEFAULT)};
-
-  &:hover {
-    background: ${({ disabled, bordered }) =>
-      !disabled &&
-      (bordered
-        ? getTypeToken(TOKENS.backgroundButtonBorderedHover)
-        : getTypeToken(TOKENS.backgroundButtonHover))};
-    color: ${({ disabled, bordered }) =>
-      !disabled &&
-      (bordered
-        ? getTypeToken(TOKENS.colorTextButtonBorderedHover)
-        : getTypeToken(TOKENS.colorTextButtonHover))}!important;
-    ${getButtonBoxShadow(BUTTON_STATES.HOVER)};
-
-    ${IconContainer} {
-      color: ${({ disabled, bordered }) =>
-        !disabled &&
-        (bordered
-          ? getTypeToken(TOKENS.colorTextButtonBorderedHover)
-          : getTypeToken(TOKENS.colorTextButtonHover))};
-    }
-  }
-
-  &:active {
-    ${({ disabled, bordered }) =>
-      !disabled &&
-      css`
-        background: ${bordered
-          ? getTypeToken(TOKENS.backgroundButtonBorderedActive)
-          : getTypeToken(TOKENS.backgroundButtonActive)};
-        color: ${bordered
-          ? getTypeToken(TOKENS.colorTextButtonBorderedActive)
-          : getTypeToken(TOKENS.colorTextButtonActive)}!important;
-        ${getButtonBoxShadow(BUTTON_STATES.ACTIVE)};
-        & ${IconContainer} {
-          color: ${bordered
-            ? getTypeToken(TOKENS.colorTextButtonBorderedActive)
-            : getTypeToken(TOKENS.colorTextButtonActive)};
-        }
-      `};
-  }
-
-  ${getFocus}
-
-  ${StyledSpinner} {
-    width: ${getSizeToken(TOKENS.loadingWidth)};
-    height: ${getSizeToken(TOKENS.loadingHeight)};
-  }
-`;
-
-StyledButton.defaultProps = {
-  theme: defaultTheme,
-};
-
-const StyledButtonContent = styled(({ theme, loading, ...props }) => <span {...props} />)`
-  visibility: ${({ loading }) => loading && "hidden"};
-  height: 100%;
-  display: flex;
-  flex-basis: 100%;
-  justify-content: center;
-  align-items: center;
-
-  // IE flexbox bug
-  ${onlyIE(css`
-    min-width: 100%;
-    max-width: 1px;
-  `)};
-`;
-
-StyledButtonContent.defaultProps = {
-  theme: defaultTheme,
-};
-
-const StyledButtonContentChildren = styled.span`
-  display: inline-block;
-`;
-
 const Button = React.forwardRef<Props, HTMLButtonElement>((props, ref) => {
   const {
-    asComponent = "button",
+    asComponent,
     children,
     bordered,
     disabled,
@@ -232,22 +43,17 @@ const Button = React.forwardRef<Props, HTMLButtonElement>((props, ref) => {
   const onlyIcon = iconLeft && !children;
   const isDisabled = loading || disabled;
 
-  warning(
-    !(!children && !title),
-    "Warning: children or title property is missing on Button. Use title property to add aria-label to be accessible for screen readers. More information https://orbit.kiwi/components/button/api/#accessibility",
-  );
-
   return (
-    <StyledButton
+    <ButtonPrimitive
       onClick={onClick}
       iconLeft={iconLeft}
       iconRight={iconRight}
       bordered={bordered}
       fullWidth={fullWidth}
       asComponent={asComponent}
+      onlyIcon={iconLeft && !children}
       disabled={isDisabled}
       loading={loading}
-      onlyIcon={onlyIcon}
       size={size}
       sizeIcon={sizeIcon}
       href={!disabled ? href : null}
@@ -256,7 +62,7 @@ const Button = React.forwardRef<Props, HTMLButtonElement>((props, ref) => {
       type={type}
       width={width}
       className={className}
-      buttonRef={ref}
+      ref={ref}
       role={role}
       circled={circled}
       submit={submit}
@@ -294,7 +100,7 @@ const Button = React.forwardRef<Props, HTMLButtonElement>((props, ref) => {
           </IconContainer>
         )}
       </StyledButtonContent>
-    </StyledButton>
+    </ButtonPrimitive>
   );
 });
