+++ src/primitives/ButtonPrimitive/components/ButtonContentChildren.js
@@ -1,0 +1,8 @@
+// @flow
+import styled from "styled-components";
+
+const StyledButtonContentChildren = styled.div`
+  display: inline-block;
+`;
+
+export default StyledButtonContentChildren;
