+++ src/primitives/ButtonPrimitive/helpers/getButtonBoxShadow.js.flow
@@ -2,7 +2,7 @@
 import type { Interpolation } from "styled-components";
 
 import type { Props, Type } from "../index";
-import type { ThemeProps } from "../../defaultTheme";
+import type { ThemeProps } from "../../../defaultTheme";
 
 type State = "default" | "focus" | "active" | "hover";
