+++ src/primitives/ButtonPrimitive/helpers/getFocus.js.flow
@@ -1,5 +1,5 @@
 // @flow
-import type { ThemeProps } from "../../defaultTheme";
+import type { ThemeProps } from "../../../defaultTheme";
 
 export type GetFocus = ({
   bordered: boolean,
