+++ src/primitives/ButtonPrimitive/helpers/getIconSpacing.js.flow
@@ -1,6 +1,6 @@
 // @flow
 import type { Size } from "../index";
-import type { ThemeProps } from "../../defaultTheme";
+import type { ThemeProps } from "../../../defaultTheme";
 
 export type GetIconSpacing = () => ({
   right: boolean,
